fixed competence link, found bug in progress
Former-commit-id: e82ebb044b901964a7d504a31dde559e12dabfab

+++ competence-gui/competence-webapp/src/uzuzjmd/competence/gui/client/progressView/ProgressEntry.java
@@ -27,7 +27,7 @@ public class ProgressEntry extends Composite {
 	interface ProgressEntryUiBinder extends UiBinder<Widget, ProgressEntry> {
 	}
 
-	public ProgressEntry(String userName, Double progress) {
+	public ProgressEntry(String userName, int progress) {
 		initWidget(uiBinder.createAndBindUi(this));
 		this.userNameLabel.setText(userName);
 		this.progressBar = new ProgressBar(Style.ANIMATED);
